Go/No-Go: Custom Plugin
This commit creates a simplified a custom copy of the jsPsychCategorizeHtmlCustom. Compared to the original, the plugin adds a "visual feedback" on the shape when the spacebar is pressed. It also removes everything not needed in the function, and comments everything else.

Another small change consists in making the messages for correct and incorrect feedback messages as global variables.